many Bug Fixes and Improvements
Fix bug when copying sweepResults Objects

Allow sweepResults plots to be saved to a given directory.

Add in functionality to shift current in double pulse results as well as withing sweep Results.

Add plot titles

Change location of default x and y samples for surface plots.

Improvements to the code that finds the starting and stopping Idxs

Spelling fixes